requirements.txt
telethon==1.32.1
python-dotenv==1.0.0
cryptg==0.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
